Fixed bug where the writehooks on quat/rotations weren't working.
Given:

    foo = new please.GraphNode();
    foo.quaternion = [0,0,0,0];
    foo.rotation = [90, 90, 90];

if you examine foo.quaternion after executing the above commands, you
will find that it is still [0,0,0,0] instead of being the quaternion
representation of the object's rotation.

Sooooooooooo

The writehook seemed to be functioning correctly.  The problem ended up
being in the getter for the group as defined in
please.make_animatable_tripple, wherein the variable "store" in the
function's scope was not equal to obj.__ani_store.  The "cache" variable
was also incorrect.  The variable "obj" however was correct, so the hack
fix was to make 'store' and 'cache' preprocessor definitions. >_>

I *really* don't like this solution.

Store would appear to be an ani_store object... just the wrong one o.O

With the 'stare' and 'cache' variables replaced with a preprocessor def,
there is no longer a 'store' variable in the function scope of the
getter (so even the wrong value isn't present either)...

I'm so confused, but it works now at least >_>