The please.load method is bypassed for default assets, so to make sure that processing happens in the desired order, it is deferred to the first please.access call.
This is accomplished by storing a function that returns the asset
instead of the asset itself into the asset cache.  Then, when
please.access is called, if the asset is a function, it calls that and
uses the result instead.  The result is then also stored to replace the
function that produced it.