Added support for nested glsl includes + fixed a fucking aweful bug.
Nested includes (for issue #222) was pretty simple to implement, but
there was a bug where sometimes but not always there would be a trailing
';' token that be left over in the compiler output that the gl driver's
compiler would barf over.

Turns out, the __identify_functions thing would prevent a semicolons
that precedes a function definition to be omitted.  Invocations in the
global scope are ignored, but their trailing semicolons are not, so some
of include statements would work fine and some of them would not.

Oh and of course I never actually docmented this behavior about the
includes being implicitly ignored, so I added a big comment for that,
and a warning next to the include macro >_>

The bug is fixed, so now the __identify_fuctions method shouldn't munch
random semicolon tokens, the include macro will complain loudly if the
semicolon is missing, and the print method has a shiv to remove them if
they follow an invocation.

>_> >_> >_>