Half-assed fix for #220 - bad macro hoisting
"It's half-assed, but it's the right half of the ass." - makes it so the
extension macro always appears before everything except the implicit
precision statements.  The should instead always appear at the very top.

Also, this just copies them to the top - the original macros should be
removed, as should redundant ones.

Good enough for now, but I want the warning to be gone before I close
this ticket.