Fix for #288 - removed vestigial code.
I previously removed the 'falloff' property from the spotlight object
because the shader no longer required it (just using inverse square
instead), but forgot to remove the driver on the light's camera object.

I considered setting it to Infinity, but that would mess up the
projection matrix.  Infinity could probably be used as the depth buffer
clear color though, but the far plane seems to work fine.