Fixed scoping problem on scalar drivers and related camera bug.
Basically in some cases that I can't reproduce, the driver property
won't use the correct objects for __ani_cache or __ani_store - the
setter function will work fine, but the getter function will return the
original vaule.  The solution was to force them to use the long path to
those objects rather than trying to scope the objects.

This fixes the problem being seen in issue #226.