Minor improvements to simple_brdf, fixed bug in orthograph cameras.
The camera bug is just the orthographic mode equivalent to the
perspective mode bug that was fixed in 5677644a.

The changes to simple_brdf.frag + related files implements what I think
is phong lighting (?) - it is a modified version of an example in a
shader programming text book, and the text did not name the equations
presented for the example >_>.  This relates to issue #226.