Ipmlemented primitive soft shadows #231
Light size is hard coded at the moment, but now 4 ray casting samples
are generated instead of just the one.

Highlights are still using the light's origin view matrix, which may or
may not be ok as is.

Some further experimentation is needed - the shadows don't seem to get
softer with depth, but idk maybe its fine?