Added a size control to spot lights (issue #231).
I don't think I'm going to continue with this approach, as the quality
of the shadows looks worse the greater the light size is.  It would need
more samples to look good, and it does seem to be not so cheap in other
ways even though it only needs one texture lookup.

It was a fun experiment, but I don't think it is what I want.