Modified JSIR constructor, started populating __drawable_ir function.
The JSIR constructor now just takes the method name and a variable
number of arguments.  If an argument is a '@', then it will be ignored,
but the next argument will be considered to be a dynamic value
regardless of if the provided default is a function or not.

The __drawable_ir function now can generate IR for updating the uniform
variables associated to the drawable.

Next up will probably be adding in buffer binds and draw calls, and
hooking this into the JTA importer.  It is likely that this output will
need to have some additional metadata so that the redundant state
changes can be removed.

Also databinding is missing.

Relates to issue #232