Added a "Signal" implementation inspired by GObject signals.
Signals are basically functions that can be given multiple bodies and
have no return value. They are intended to be used for event
dispatching.

A Signal object can be called like a function (because it is one), but
you must attach callbacks to it to provide it's behavior. The
"represented" argument is the 'this' value for the callback methods. If
"represented" is missing or is null, then 'this' will be the Window
object.