Static draw functions now automatically recompiled when necessary.
All shader bindings for GraphNodes now have a common signal to use as a
write hook;  the draw function's intermediary representation now listens
to this, and when the corresponding property is written to, the IR
updates.  When becomes dynamic, it calls another signal, which triggers
the recompile.

In addition to this, uniform properties are static first, and then add
cache entries only when they've been updated.

The only exception is propertise that start out as driver properties,
but in the future, there will be some kind of "make static" thing to
force everything to be static.  This would be as part of a greater
mechanism for bypassing the scene graph entirely.

This relates to issue #232.