Monolithic static drawing on the SceneGraph objects #232.
Now, objects which support static drawing do not provide their own draw
functions (static or otherwise).  Instead, they provide IR, and signal
to their GraphRoot when it has changed (when applicable).

Adding and removing objects from the graph will also update the static
draw function.

It can be signaled many times, but setTimeout is used to make sure it is
recomplied out of phase with the current callstack, and so redundant
calls are eliminated.

Performance seems marginally better so far than the old way in some
demos, but there are a lot of ways this can be optimized further.  One
thing that needs to change is to only mark IR corresponing to uniforms
as dynamic if it was updated *after* the first compile.