Bugfix / hack for code generation in issue #232
This makes it so that if you set an "=" method IR to set something to
null, then it'll just return an empty string while compiling.

It may be the case that this needs to be more specified so that it only
does this special case for shader vars, but that is all "=" IR is used
for at this time.