Graph drawing functionality moved from m.graph to m.compositing.
This commit moves the graph drawing code off of the SceneGraph object,
and into the RenderNode objects.  Sort of.  This also takes the entire
function and splits it apart into a prefix and postfix IR for the
generated static draw function, resulting in the entire draw function
being generated code.

Otherwise, this should be more or less equivalent to the old way of
doing things, except that by associating the draw function to
RenderNodes, we can assume a 1:1 relationship between shaders and
RenderNodes, but a 1:many relationship between SceneGraph instances and
RenderNodes.

As this commit relies on a yet-fictitious "quote" macro, this is almost
certainly broken code.

Relates to issue #232.