wrapped node ir list in a object
The one-off drawable ir list on JTA instances has been removed, and
replaced with a functionally equivalent object.  Various files have been
updated to fit this new scheme.

The IR list is now accessed by calling the 'generate' method on the
DrawableIR object.  This should make it simpler to produce program
specific IR, but still allow for some state tracking for the sake of
guessing when properties should be considered static.

This fuzzies up what "IR" is exactly in this sense though, so that is
annoying.  I think this will need to be tinkered with further.

Relates to issue #232.