Added GraphNode.freeze method.
Calling the 'freeze' method on a GraphNode which employes DrawableIR
will cause all of the bindings to be marked static and coerce function
bindings to cached return values.

This is offered in replacement to the "use_manual_cache_invalidation",
and provides no reverse method.  It is vastly superior in performance,
but is premenant, so use with care.

This pertains to issue #232.