Refactored event dispatching for picking events.
Moved code specific to detecting click and double click events into
m.picking.js.  Event stubs on GraphNodes are now signals.  The API can
be mostly used as before, though it does not allow for event stubs to be
arrays anymore.

The Object Selection demo has been updated to be consistent with the new
API changes and best practice.

Pertains to issue #237.