Updated the picking distortion API
Now you provide a function that is responsible for applying the
distortion to the picking coordinates.  This does mean that you can't
just re-use the distortion shader, though this is structured to make it
easy to port.

The benefit of this, is that it simplifies the picking system quite a
bit, and makes your game run faster.

The Distorted Selection demo uses this new API.

Pertains to issue #237.