Added frame limiting to mouse move picking events.
Experimental feature - makes the actual frame rate better the lower the
picking rate is set, unsurprisingly.  However, the low sampling rate
makes it look like the frame rate is in fact lower... unsurprisingly.

I'm checking this in because I think it will still be useful.  One of
the following should probably happen:

 1. the sampling rate is kept low, but the full picking buffer would be
    saved for move events - this would then be used to sampling between
    sampling frames.

 2. possibly restrict move events to never pick location

 3. a setting be exposed for the sampling rate of move events

This pertains to issue #237.