Meaningful sourceURL directives in generated code.
Added sourceURL directive to generated functions.  This allows for
firefox to recognize these sources in the debugger, making it easier to
debug them.

The url is composed of the __id of the graph node that it describes and
the name of the shader that renders it.  It is formatted like so:

 > "rendernode:uuid:shadername.js"

This link has information about the sourceURL directive:
https://developer.mozilla.org/en-US/docs/Tools/Debugger/How_to/Debug_eval_sources