corrected object index assignment for legacy nodes
This change ensures that node.shader.object_index is set durring the
picking pass.  The previous version did not work, probably because it
was setting the value assuming a driver rather than writing to the
ani_store.

I also changed it to use __states and __alpha instead of __flat.  I
don't think this was strictly necessary, but it reflects the logic in
the generated rendering function, and also probably means __flat can be
removed.