Removed the buggy 8x8 picking buffer stuff.
This commit removes the premature optimization inspired by tiled
rendering, wherein the picking passes would only use an 8x8 rendering
buffer.  While it seemed to work fine for most cases, a large quad would
occasionally not render, and I'm still not sure why... but reverting
this fixes the problem, and the performance gain was minimal anyway.

This commit also adds a condition to the skip_when function for the
picking_pass frame handler, so as not to blow up when
please.__compositing_viewport is null.

This commit relates to issue #237.