Hack which fixes location picking.
I'm not sure exactly if this is the best solution, as I'm pretty sure it
cheeses any rate limiting on picking -- at least, if it is enforced via
the compositor.  If it is not, then this is a good solution, but should
be perhaps given a formal option in the call for rendering, or as a
property of the RenderNode.

Anyways, might be prudent to audit this and make sure it isn't doing
anything silly now like spawning tons of indirect buffers or something.

Relates to issue #237.