"Graphless" IR generation for JTA / stamping API.
I went and implemented an API for generating static drawable IR.  The
issue notes refer to this as "graphless" drawing, but using a singular
GraphNode as a "cursor" ended up making this really trivial to
implement, so I went and did that.

The API works like this:

```js
var tracker = please.access("object.jta").instance();
for (var x = -5; x<=5; x+=1) {
    tracker.location_x = x;
    graph.stamp(tracker);
}
```

This commit relates to issue #232.