Fixing bad behavior in stamp API
Problem was, the cursor object caches values for its getter / setters,
and we want to bypass that.  This is a partial fix, though.

In this case, when it reads out the value of a driver, it is able to
bypass the cache by calling the driver directly.  This does not work if
the driver references another driver, though.

What this should probably do instead is just delete the GraphNode object's
ani cache before each call to ```__DrawableIR::copy_freeze```.

Relates to issue #232.