Experimentally added support for heap to compiled draw ir.
Specifically, only for stamped IR right now.

I'm checking this into a new branch for reference, but this is currently
a failed experiment, as it makes the static drawing demo run *slower*.
I'm not sure why, as I would expect them to have equivalent GC churn,
but apparently not.

I think it could run faster if the "new Float32Array" stuff was passed
directly into a webgl call, in which case it would likely function as
casting a pointer.  But since it is instead being passed into
boilerplate, it likely creates a new object, and thus runs like garbage.

The solution to this is probably to reduce the JSIR objects down to the
webgl calls.

This relates to issue #232.