Disabled cache checks on old uniform boilerplate.
The cache still exists and is written to, but it does not prevent
redundant uniform uploads anymore.

Bypassing the boilerplate setters can lead to the value of the cache to
be incorrect, preventing uploads through the boilerplate that should
actually happen.  The cache is much less important now, since we already
prune redundant calls through other means.

This pertains to issue #238 and fixes the remaining issue with the
"models" demo.