Gani graph instancing works w/ webgl again.
This is a bit of a hack.  The please.gani.build_gl_buffers call ideally
would be called when an please.media.__AnimationData object is
constructed, but it also needs for the images referenced by the attrs to
be downloaded first.

So, the hack is to just defer this until the GraphNode's draw method is
called.  please.gani.build_gl_buffers will now return early instead of
raise an assertion error.  If no vbo is present, then nothing is drawn.

A better solution would involve having a better signalling system for
when prerequisite assets have finished downloading.  Should also
consider having some kind of place holder gani.  These issues can be
addressed later, though.

This relates to issue #240.