Unbind textures after draw pass, better debugging.
This commit covers two things.  The first relates to issue #244, wherein
at the end of a generated render function, M.GRL will now unbind all
textures that were bound by the shader.  This is to avoid interfering
with framebuffer attachments, though this doesn't seem to entirely solve
the problem.

The second change is the addition of a DEBUG preprocessor definition.
When enabled, it forces gl debugging to be on, and adds some additional
runtime assertions.  I'm currently using this to help track down the
problem in #244, but I think this will be good overall as another QA
tool.