Removed render texture feedback loop, improved debug mode.
This commit introduces a hack to force all texture units to unbind
before setting the frame buffer to a texture.  This should effectively
prevent some texture binding feedback loops and undefined behavior,
though it is possible to bind the texture after calling set_framebuffer,
so it is kind of on the user to not do that.

If m.grl was compiled with DEBUG true, then this will catch case as an
error anyway.

I was hoping this would solve #244, but that does not seem to be the
case.  Still, good to address errors and warnings first anyway.