Texture arrays, more debugging tools, vast refactoring.
This is a somewhat large update, which covers several things:

 - Added DebugViewportSplitter RenderNode.  This essentially splits the
   screen into up to 9 viewports, drawn above whatever you normally
   would be drawing.

 - Added FloatingPointBufferViewer RenderNode.  This is needed to do
   debug rendering of floating point textures.

 - Added support for uniform arrays of textures!  The debug viewport
   splitter uses this.

 - DeferredRenderer uses correct framebuffer id for light passes again.

 - Activating a shader no longer copies over uniform values from
   previous shader.  Instead it is expected that one will use
   RenderNodes or GraphNodes to set shader variables.  I also added a
   means in which M.GRL's few global uniform variables can be
   persistent, but it is intentinoally not exposed as a generic
   interface.  The reason for avoiding globals is so that we don't get
   weird bugs when shaders define common names (like "enabled") but with
   incompatible types.  This broke virually nothing, so I'm not
   considering this to be backwards-incompatible.

This relates to issues #232, #244, and #247.

There is a regression with picking in the bezier demos, where the
drag-and-dropping causes objects to shoot off to one side.  Normal
picking demos seem to work ok still, though.

Also, FloatingPointBufferViewer curiously works fine when used with
PictureInPicture (even several nested ones), but not in
DebugViewportSplitter.  In the later case, they just render completely
blank, which is similar to issue #244!  I am unsure why this is!