Basic syntax support for instancing globals.
This commit implements support for recognizing instancing
globals (identified by the "in/uniform" identifier), as well as
generating "virtual globals" as a result.

The ast "Global" recognizes "in/uniform" as a valid mode type, and the
constructor will automatically generate all of the virtual globals
needed for printing valid GLSL.

A macro will be needed in the beginning of the main method to populate
the non-uniform non-attribute global generated for the scripts to
reference.

Mat3 and mat4 variables are split into several attributes, and will
require some additional support code for reassembly.

This does not include the necessary boilerplate to use these variables
generically.

Relates to issue #197.