Corrected boilerplate for instancing matricies.
When using the "in/uniform" syntax for instancing, the special cases for
matrix types should be handled now.

For mat2, this means that the attribute type will be vec4, and the
boilerplate will use a mat4 constructor and swizzle syntax.

For mat3 and mat4, the attribute types will be vec3 and vec4, so that
the path to generating the boilerplate constructor can be shared.

Relates to issue #197.