Compiler can now detect when stamps are instanceable.
This changes how stamps are tracked by m.graph.js, so that they are
sorted into a state group (defined by the vbo/ibo used).  Variance in
"defaults" is tracked.  A dictionary is used to track what hasn't
changed yet.  When something contradicts an earlier entry, it is removed
from that dictionary and then added to the appropriate __ir_variance
list.

The draw call compiler now is able to see when a vbo/ibo pair is a good
candidate for instancing based on the number of objects to be drawn, but
also on the basis of if the applicable uniforms in __ir_variance are
actually instancable.

This change does NOT introduce code for managing the instancing buffers,
and the place where the instancing codegen would be inserted is just
stubbed out, so this should not actually cause anything to run
differently besides a bit of extra book keeping.

This change also exposes the "instanceable" object on shader program
objects, and marks the world matrix in various bundled shaders as being
instanceable.

This relates to issue #246.