Hack implementation of automatic instancing for stamps.
There are some hacks that need to be cleaned up.  This might amount to
refactoring some things in m.buffers.js so as to provide a unified way
to generate IR for instanced drawing w/ two or more different buffer
objects.

A short TODO list of things that should be absolutely cleaned up:

1. If a recompile event were to be trigger later on, this would result
   in the creation of a new instancing buffer.  Instead, we should
   either reuse the existing if possible.

2. Need an not-hardcoded way to enable the instance control uniforms.
   AS IS, THIS WILL CURRENTLY ONLY WORK FOR world_matrix.

3. Better teardown - use the state tracker properly for disabling the
   attributes, and turning off instancing etc.

4. Test instancing multiple things + rendering non-instanced things.

5. Generated comments to denote generated instancing code.

Relates to issue #246.