Generalized hardcoded instancing enabler.
This removes the hardcoded hack to enable/disable instancing for
world_matrix, and instead incorperates it into __DrawableIR.generate.

So now, the enabler uni should have proper state tracking, the
corresponding uniform should not be set when instancing, and it should
also be disable automatically whatever follows, assuming it is not also
instanced.

This relates to issue #246.