Non-instance buffers can be bound w/ instanced buffers.
This is used to reduce state churn and hacks.  When this is used, the
instanced buffer will be bound normally.  Then, the non-instanced buffer
will bind as it normally would, but it wont disable attributes used by
the instanced buffer.

This also removes the "instances" argument from several methods in
favor of tracking that on the instanced buffers themselves.

This relates to issue #246.