Added caching / deletion mechanism for instance buffers.
This adds the "destroy" method to vbo containers.  When called, it
should de-allocate the buffer object, and mark itself as being deleted.
In debug builds, attempting to call the bind or draw methods will raise
an error after the buffer has been deleted.

This also adds a caching system for the stamp instancing buffer + cache
invalidation mechanism, though it hasn't been properly tested at the
time of writing.

This relates to issue #246.