Added make_grid method for making a grid of quads.
This change adds the method please.gl.make_grid, which takes a tile
width, height, and the number of times it should be repeated along the X
and Y axis, as well as an optional origin.  This produces a VBO
containing the tesselated quad described by the arguments.  This is
mainly intended to be distorted by the vertex shader to produce effects
such as height maps and waves.

Additionally, this commit removes the old logic for please.gl.make_quad,
as it has been updated to call please.gl.make_grid.

This relates to issue #64.