Swappable functions bugfixes.
This commit fixes a compiler error complaining about attempting to
return a value when the function's return type is void.  Now, if a
swappable function has a void return type, the rewrite will not attempt
to return the result of the plugin functions called.

This commit also resolves an issue where parameter positions in input
lists were assumed to be a fixed value.  The code now uses a relative
index via slicing, to allow qualifiers.