Added vertex animation waves to glsl compiler demo.
This makes use of the new please.make_grid method to generate tesselated
surfaces.  This change adds a new vertex shader part, "ext.vert" which
is used in a similar way to "ext.frag", though it is not current exposed
to the user for editing.

Vertex animation is achieved by way of the "morph" swappable function.

Unfortunately, the procedural texture of the water is just rendering
black.  This is because the normals are bogus and cause the water
texture to render a black surface.  So, that needs to be fixed yet.

Relates to issue #64.