Normals generated correctly now for generated grids.
This resolves issue #255.  The core issue was a typo, that caused the
array for the generated normals to be bound to "noraml" not "normal, and
also that the local_normal was not being used to generate the
world_normal in simple.vert.

In addition to correcting both of those, this commit also enhances the
demo by calculating a new surface normal to match the wave form.