Global attribute cache fixes picking problem in some demos.
This change also updates the debug builds to flush after each command,
so that gl errors propogate right away.

This change updates the attribute enabling/disabling api, so that
results are persistent across shaders, rather than local to shader
bindings.  Otherwise, the behavior differs on some platforms, causing
arrays to be bound when they shouldn't be or not bound when the should
be.

This relates to issue #247, in that it resolves the problem with the
dynamic_glsl demo, but not the picking demos :/