Cleaning up technical debt in GLSL->GLSL compiler.
This commit is an attempt to clean up the horrible technical debt that
has accumulated in the print_program method in the glsl compiler.

And by that, I mean, I rewrote the function from scratch.

This isn't a perfect replacement, and there are currently some
regressions which are evident by running through some of the demos, but
this is a much clearer approach, and I think will make further
development on the compiler much more pleasant.

This relates to issues #260 and #261.